added script to generate a markdown reference for all commands, added "all" support for !reloadmod, moved commandconsole.py to util folder

+++ build.py
@@ -1,10 +1,11 @@
 from os import system
 from pathlib import Path
 from shutil import rmtree
+from sys import executable
 
-build = Path('build/')
-dist = Path('dist/')
-egg = Path('PythonTwitchBotFramework.egg-info/')
+build = Path('build')
+dist = Path('dist')
+egg = Path('PythonTwitchBotFramework.egg-info')
 
 
 def delete_build_folders():
@@ -17,6 +18,7 @@ def delete_build_folders():
 
 
 delete_build_folders()
-system('py setup.py sdist bdist_wheel')
+system(f'{executable} setup.py sdist bdist_wheel')
 system('twine upload dist/*')
-delete_build_folders()
+if input('remove build folders? [Y/N]: ').lower().strip() == 'y':
+    delete_build_folders()
